requirements.txt
discord.py>=2.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
paramiko>=3.0.0
PyJWT>=2.8.0
